pyproject.toml
[tool.poetry]
name = "crucible"
version = "0.1.0"
description = "Develop better LLM apps by testing different models and prompts in bulk."
authors = ["Gustavo Araujo Costa <gustavo@arcos.org.br>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
anthropic = "^0.30.1"
openai = "^1.35.11"
tiktoken = "^0.7.0"
python-dotenv = "^1.0.1"
ollama = "^0.2.1"
streamlit = "^1.37.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
rich = "^13.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
